requirements.txt
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.3.0
idna==2.10
pyaml==23.9.7
pycodestyle==2.6.0
PyYAML==6.0.1
requests==2.31.0
toml==0.10.2
urllib3==1.25.11
